requirements.txt
asgiref==3.8.1
Django==5.1.3
psycopg2-binary==2.9.10
python-decouple==3.8
djangorestframework==3.15.2
django-cors-headers==4.6.0